HB-7996: Sleep 1s instead of 15m in RetryRunner to ensure TERM signal handling occurs.
RetryRunner sleeps for 15 minutes in _snooze. When the TERM signal
comes in, it interrupts this sleep but then restarts it and allows
it to finish before the signal handler gets called.

So:

If the sleep time is something very short, like 1 second, the
signal handling happens reasonably promptly, and the process will
exit as expected.

If the sleep time is something medium length, like 20 seconds,
then we might have to wait up to 20 seconds (depending on how far
into the sleep we were) before the signal handler is called.

If the sleep time exceeds 90 seconds, as it did prior to this change,
then we will run into systemd's own 90 second timeout, at which
point it will switch from TERM to KILL to circumvent this problem.

It's normal in mailman for a qrunner process to have a sleep time
of 1 second, so we might as well do the same thing with RetryRunner.