Fix byte prefix when reading header values from emails
Case HB-8030

When pulling header values out of an email, we were encoding them to a
character set. We just pass along the string now as it should be UTF-8
already.

It's possible this could break something else that calls this code. But
all of the callers expect a normal string, so it should at least be
better than before.

Changelog: