Fix empty attachment and raw byte emails.
Case HB-8042

Depending on how the decode of certain strings failed two outcomes could
happen. The email could get wrongly marked as a multipart email with an
empty attachment, or it would output raw byte strings as the message
body. Both of these issues are fixes with this commit.

Changelog: