HB-8147: Fix oversize message detection for messages with attachments.
In earlier fixups, we replaced usage of email.Iterators.body_line_iterator()
with msg.get_payload() in the size checking of Hold.py because
body_line_iterator was apparently missing in python 3, but get_payload only
worked for non-multipart messages. As soon as the message had an attachment
(or was multipart for some other reason), this oversize message detection
would fail because get_payload returns None for anything multipart. Rather
than implement the body_line_iterator behavior ourselves to traverse
multipart messages, it turns out we can go back to using body_line_iterator
as originally intended; we just need to use lowercase "iterators" because
that's the name of it in python 3.