Fix attachments saved to public and private archives.
Case HB-8149

For public archives, the attachment needed to be treated as a byte
string when saved to the disk.

For private archvies, the cgi script was trying to print binary data to
stdout which was redirected to a StringIO object. StringIO lacks methods
to deal with binary stings. To fix this we point stdout back to the
"normal" stdout which can deal with binary strings. Once we are done, we
move stdout back to the StringIO version.

Changelog: