Fix membership list links when using pagination.
Case HB-8107

When the Mailman UI wants to paginate the membership list it creates
buckets of users based on the first letter of the email address. The
user can click on a link in the member list table to see the email
addresses that begin with that letter. These links were being encoded into
a byte string. That no longer needs to happen.

Changelog: Fix membership list links when using pagination.